# packages/api-contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: iF Fleet Management API
  version: "0.1.0"
  description: |
    Single backend API for the iF Fleet Management Platform.
    Supports Employee, Driver, and Admin roles on web and mobile.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://staging-fleet-api.internal/api/v1
    description: In-house Docker staging
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message: { type: string }
        errors: { type: array, items: { type: string } }
...
    RequestOtpBody:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Must be company-domain email
    VerifyOtpBody:
      type: object
      required: [email, employeeId, otp]
      properties:
        email: { type: string, format: email }
        employeeId: { type: string }
        otp: { type: string, minLength: 6, maxLength: 6 }
    AuthTokens:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            email: { type: string }
            role: { type: string, enum: [EMPLOYEE, DRIVER, ADMIN] }

    # ── Preset Location ─────────────────────────────────────────────────────────
...
paths:
...
  /auth/request-otp:
    post:
      tags: [Auth]
      summary: Request OTP for login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RequestOtpBody' }
      responses:
        '200':
          description: OTP sent to provided email
        '400':
          description: Invalid email or non-company domain
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP and employee ID; returns JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyOtpBody' }
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthTokens' }
        '401':
          description: Invalid OTP, expired OTP, or unmatched employee ID
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: New access token issued
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke current session
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Session revoked

  # Preset Locations
...